requirements.txt
sqlalchemy~=2.0.48
alembic~=1.18.4
fastapi~=0.135.2
passlib~=1.7.4
python-dotenv~=1.2.2
sqlalchemy-utils~=0.42.1
pydantic~=2.12.5